Gen ai Project/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
streamlit
streamlit-option-menu
pinecone
sentence-transformers
scikit-learn
pandas
numpy
